[CLEANUP][DB] Misc cleanup changes for DB support - Change all ID usages to Long - Retrieve DbObject from dbId wherever required to fix errors due to changes in Objects - Use dbId instead of old id for Lead - Use SyncDbTask Instead of GetTasks whereever necessary - Add RefreshTasks static API to homescreen and use everywhere - Add static API to FormField to use for converting from JSonArray to Object - Update Homescreen DB Sync and UI Init mechanism - Add Sync button to homescreen - Remove Static APIs which were earlier storing Tasks DB - Add Form Saving & Task closing support to SubmitFormTask. Form should be saved after retrieving id and version from server

+++ app/src/main/java/com/biz/navimate/activities/BaseActivity.java
@@ -3,7 +3,6 @@ package com.biz.navimate.activities;
 import android.Manifest;
 import android.app.Activity;
 import android.content.Intent;
-import android.graphics.Bitmap;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.v4.app.ActivityCompat;
@@ -355,7 +354,7 @@ public abstract class BaseActivity extends AppCompatActivity {
                 {
                     if (resumeResultCode == Activity.RESULT_OK)
                     {
-                        leadPickerListener.onLeadPicked((ArrayList<Integer>) resumeResultIntent.getSerializableExtra(Constants.Extras.LEAD_PICKER));
+                        leadPickerListener.onLeadPicked((ArrayList<Long>) resumeResultIntent.getSerializableExtra(Constants.Extras.LEAD_PICKER));
                     }
                 }
                 break;
